Don't remove the unix socket if we weren't the one who created it.  (In other words, clients should not remove the socket file.  Duh.)
git-svn-id: svn://svn.freevo.org/kaa/trunk/base@3184 a8f5125c-1e01-0410-8897-facf34644b8e

+++ src/notifier/sockets.py
@@ -469,7 +469,7 @@ class Socket(object):
         self._queue_close = False
 
         self._socket.close()
-        if isinstance(self._addr, basestring) and self._addr.startswith('/'):
+        if self._listening and isinstance(self._addr, basestring) and self._addr.startswith('/'):
             # Remove unix socket if it exists.
             try:
                 os.unlink(self._addr)
